mqtt_cli_python/requirements.txt
paho-mqtt>=2.0.0
asyncio>=3.4.3
